requirements.txt
django==3.0
flake8==3.7.0
pytest-django==3.9.0
pytest-cov==5.0.0